add user entity creation on first login
