+++ app/lib/actions.ts
@@ -3,6 +3,8 @@
 import {signIn} from '@/../auth'
 import {Pages} from '@/constants/Pages'
 import {createSong, fetchSongById, updateSong} from '@/lib/data/songs/actions'
+import {QueryResult, QueryResultRow, sql} from '@vercel/postgres'
+import type {User} from 'next-auth'
 import {AuthError} from 'next-auth'
 
 export {createSong, updateSong, fetchSongById}
@@ -19,4 +21,17 @@ export async function authenticateGoogle() {
     }
     throw error
   }
+}
+
+export async function checkUserExists(user: User): Promise<boolean> {
+  const email = user.email
+  let result = await sql`SELECT EXISTS ( SELECT 1 FROM users WHERE email = ${email} ) AS user_exists;`
+  return result.rows[0].user_exists
+}
+
+export async function createUser(user: User): Promise<QueryResult<QueryResultRow>> {
+  const id = crypto.randomUUID()
+  const password = crypto.randomUUID()
+  const provider = 'Google'
+  return await sql`INSERT INTO users (Id, Name, Email, Password, Provider) VALUES (${id}, ${user.name}, ${user.email}, ${password}, ${provider});`
 }
