+++ middleware.ts
@@ -1,5 +1,5 @@
+import {authConfig} from '@/../auth'
 import NextAuth from 'next-auth'
-import {authConfig} from './auth.config'
 
 export default NextAuth(authConfig).auth
